fix bug - entering whole numbers with 10 or more digits causes a "For input string: XXXXXXXXXX" Exception

+++ src/calculator/Calculator.java
@@ -1,8 +1,9 @@
 // TODO: add keyboard input
 // TODO: experiment increasing the font size on buttons and display
 // TODO: attempting to overflow the width of the main display resizes the program in an undesirable way
-// TODO: bug - entering whole numbers with 10 or more digits causes a "For input string: XXXXXXXXXX" Exception
 // TODO: possibly use Backspace character on delete button
+// TODO: experiment with light gray background color for main display
+// TODO: experiment with scientific notation for very large or very small numbers
 
 /*
  * To change this license header, choose License Headers in Project Properties.
@@ -323,7 +324,7 @@ public class Calculator extends JFrame {
                 
                 /* Prevent new expressions from prepending a 0 after evaluating
                 an expression that results in 0. */
-                if (Integer.parseInt(result) != 0) {
+                if (Double.parseDouble(result) != 0) {
                     printResult(result);
                 }
